chore: Export custom themes table from main `~/db/schema.ts` file in case the same situation happens
- We load the custom theme right after loading the custom fonts.
