+++ mobile/src/modules/customization/theme/utils.ts
@@ -1,7 +1,7 @@
 import { Uniwind } from "uniwind";
 import { Appearance } from "react-native";
 
-import type { customThemes } from "./core/schema";
+import type { customThemes } from "~/db/schema";
 
 import type { CustomTheme, DefaultTheme, HexColor } from "./core/constants";
 import { DefaultThemeOptions } from "./core/constants";
